[v2] refactor:  게시글 상세페이지(댓글 수정, 댓글 조회) 타입 및 훅 분리
