+++ 03_trip-trip/03_cna-v2/src/components/boards-detail/comment-write/types.ts
@@ -1,0 +1,3 @@
+export interface CommentWriteProps {
+  boardId: string
+}
